refactor config and listener package
