+++ elastic-job-lite-tracing/elastic-job-lite-tracing-api/src/test/java/org/apache/shardingsphere/elasticjob/lite/tracing/event/JobExecutionEventTest.java
@@ -15,9 +15,8 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.elasticjob.lite.tracing;
+package org.apache.shardingsphere.elasticjob.lite.tracing.event;
 
-import org.apache.shardingsphere.elasticjob.lite.tracing.event.JobExecutionEvent;
 import org.hamcrest.CoreMatchers;
 import org.junit.Test;
